refactor(binder,cli,fourslash): split over-cap files to satisfy §19
Split three hand-authored Rust source files that exceeded the 2000-line
hard limit, and fix a missing path require in the fourslash worker.

Rust splits (no functional change):
- binding.rs (2001 → 1869): extract 5 scope management methods to
  binding_scope.rs (140 lines)
- state/core.rs (2255 → 1865): extract 9 JSDoc @import tag parsing
  methods to core_jsdoc.rs (398 lines)
- handlers_code_fixes.rs (3928 → 1881): extract fallback helpers to
  handlers_code_fixes_fallbacks.rs (1458 lines) and synthetic diagnostic
  helpers to handlers_code_fixes_synthetic.rs (622 lines)

Fourslash fix: add require("path") to test-worker-patch-test-state.cjs,
which used path.basename() on 11 call sites after the worker split.
