+++ crates/tsz-cli/src/bin/tsz_server/main.rs
@@ -32,9 +32,11 @@
 mod check;
 mod handlers_code_fixes;
 mod handlers_code_fixes_enum_member;
+mod handlers_code_fixes_fallbacks;
 mod handlers_code_fixes_implement_interface;
 mod handlers_code_fixes_imports;
 mod handlers_code_fixes_jsdoc;
+mod handlers_code_fixes_synthetic;
 mod handlers_code_fixes_utils;
 mod handlers_completions;
 mod handlers_completions_display;
